# apps/backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "EuroStrip API",
        "version": "0.1.0",
        "description": "EuroStrip backend API"
    },
    "servers": [
        {
            "url": "http:\/\/localhost:8000\/api"
        }
    ],
    "paths": {
...
        "\/gateway\/commands": {
            "post": {
                "operationId": "gateway.commands",
                "summary": "Queue a protocol command for the connected EuroScope plugin",
                "tags": [
                    "Console"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/EnqueuePluginCommandRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "integer",
                                    "const": 202
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#\/components\/responses\/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#\/components\/responses\/ValidationException"
                    },
                    "403": {
                        "$ref": "#\/components\/responses\/AuthorizationException"
                    }
                }
            }
        },
...
        "\/euroscope\/messages": {
            "post": {
                "operationId": "euroscope.messages",
                "summary": "Ingest a batch of protocol messages from the EuroScope plugin",
                "tags": [
                    "PluginTransport"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/RecordPluginMessagesRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "integer",
                                    "const": 204
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "An error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#\/components\/responses\/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#\/components\/responses\/ValidationException"
                    },
                    "403": {
                        "$ref": "#\/components\/responses\/AuthorizationException"
                    }
                }
            }
        },
...
    },
...
}